@@ -1,3 +1,3 @@
 Spree::Product.class_eval do
-  has_many :videos, :order => 'position ASC', :as => :watchable
+  has_many :videos, -> { order 'position ASC' }, :as => :watchable
 end
